refactor(review-hub): decouple state from modal chrome
- Extract ReviewHubContent with all staging state, IPC, and interaction handlers
  so the body can mount in a non-modal container without dragging createPortal /
  useOverlayState along with it
- Pull pure types/helpers into reviewHubUtils.ts (PushErrorState, banner config
  table, push-banner helpers, file filter/sort utilities)
- ReviewHub.tsx becomes the modal shell only: backdrop, portal, focus init,
  overlay registration; defers everything else to ReviewHubContent
- Preserves the atomic isOpen→reset block (#4958), commitMessage immune to
  polling re-renders (#4220), and worktree.onUpdate subscription disposal (#4754)
